scraping_and_Indexing/requirements.txt
selenium
requests
boto3
python-dotenv
langchain
python-docx
openai
tiktoken
faiss-cpu
pinecone-client
flask
